+++ src/main/java/com/gachifarm/dao/AccountDao.java
@@ -1,0 +1,5 @@
+package com.gachifarm.dao;
+
+public class AccountDao {
+
+}
