+++ src/main/java/com/gachifarm/domain/Account.java
@@ -1,13 +1,14 @@
 package com.gachifarm.domain;
 
-public class Account {
+import java.io.Serializable;
+
+@SuppressWarnings("serial")
+public class Account implements Serializable{
 	String user_id;
 	String password;
 	String userName;
 	String phone;
 	String email;
-	String addr1, addr2;
-	String zipCode;
 	
 	public String getUser_id() {
 		return user_id;
@@ -39,31 +40,13 @@ public class Account {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public String getAddr1() {
-		return addr1;
-	}
-	public void setAddr1(String addr1) {
-		this.addr1 = addr1;
-	}
-	public String getAddr2() {
-		return addr2;
-	}
-	public void setAddr2(String addr2) {
-		this.addr2 = addr2;
-	}
-	public String getZipCode() {
-		return zipCode;
-	}
-	public void setZipCode(String zipCode) {
-		this.zipCode = zipCode;
-	}
 	
 	
 	@Override
 	public String toString() {
-		return "Account [user_id=" + user_id + ", password=" + password + ", userName=" + userName
-				+ ", phone=" + phone + ", email=" + email+ ", addr1=" + addr1+ ", addr2=" + addr2
-				+ ", zipCode=" + zipCode + "]";
+		return "Account [user_id=" + user_id + ", password=" + password + ", userName=" + userName + ", phone=" + phone
+				+ ", email=" + email + "]";
 	}
+	
 	
 }
